pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sinan-agentic-core"
version = "0.7.0"
description = "State-of-the-art framework for building AI agents with OpenAI Agents SDK"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Thomas Pernet-Coudrier", email = "t.pernetcoudrier@gmail.com"}
]
keywords = [
    "ai",
    "agents",
    "llm",
    "openai",
    "multi-agent",
    "orchestration",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    "openai-agents>=0.8.0",  # OpenAI Agents SDK
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
mcp = [
    "mcp>=1.0.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "mcp>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/thomaspernet/sinan-agentic"
Documentation = "https://github.com/thomaspernet/sinan-agentic#readme"
Repository = "https://github.com/thomaspernet/sinan-agentic"
Issues = "https://github.com/thomaspernet/sinan-agentic/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["sinan_agentic_core*"]
exclude = ["workflows_core*", "tests*"]

[tool.setuptools.package-data]
sinan_agentic_core = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=sinan_agentic_core --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["sinan_agentic_core"]
omit = [
    "sinan_agentic_core/__init__.py",
    "sinan_agentic_core/agents/__init__.py",
    "sinan_agentic_core/agents/orchestration/__init__.py",
    "sinan_agentic_core/guardrails/__init__.py",
    "sinan_agentic_core/tools/__init__.py",
    "sinan_agentic_core/utils/__init__.py",
]

[tool.coverage.report]
show_missing = true
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
]
